requirements.txt
requests>=2.31.0
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.14.0
dbt-bigquery>=1.7.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
chromadb>=0.4.0
streamlit>=1.30.0
pyyaml>=6.0
google-generativeai>=0.5.0
